training/requirements.txt
torch>=1.7.1
torchvision>=0.8.2
torchaudio>=0.7.2
numpy>=1.19.2
Pillow>=8.0.1
dominate>=2.6.0
visdom>=0.1.8.9
kaggle>=1.5.12
matplotlib>=3.3.3
tqdm>=4.54.1
mlflow>=1.12.1
scikit-learn>=0.24.0
tensorboard>=2.4.0
ipython>=7.19.0
python-dotenv>=0.19.0 